+++ domain/src/main/java/com/multimoney/domain/model/accountsmart/GlobalRequest.kt
@@ -1,7 +1,7 @@
 package com.multimoney.domain.model.accountsmart
 
 data class GlobalRequest(
-    val idGlobalRequest: Int,
+    val idGlobalRequest: Long,
     val accountExists: Boolean,
     val idSysRequest: Int
 )
